# internal/handlers/enrollments.go
package handlers

import (
	"attendance-management/internal/models"
	"attendance-management/internal/services"
	"database/sql"
	"encoding/json"
	"log"
	"net/http"
)

type EnrollmentHandler struct {
	Service *services.EnrollmentService
}

func NewEnrollmentHandler(s *services.EnrollmentService) *EnrollmentHandler {
	return &EnrollmentHandler{Service: s}
}

func (h *EnrollmentHandler) EnrollmentHandler(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.GetEnrollments(w, r)
	case http.MethodPost:
		h.CreateEnrollment(w, r)
	default:
		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
	}
}

// GetEnrollments 전체 수강신청 목록을 조회합니다
// @Summary 전체 수강신청 목록 조회
// @Description 모든 수강신청 목록을 조회합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Success 200 {array} models.Enrollment
// @Failure 500 {string} string "Internal Server Error"
// @Router /enrollments [get]
func (h *EnrollmentHandler) GetEnrollments(w http.ResponseWriter, r *http.Request) {
	enrollments, err := h.Service.GetAllEnrollments()
	if err != nil {
		log.Println("등록 조회 실패:", err)
		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		return
	}
	writeJSON(w, http.StatusOK, enrollments)
}

// CreateEnrollment 새로운 수강신청을 등록합니다
// @Summary 수강신청 등록
// @Description 새로운 수강신청을 등록합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Param enrollment body models.Enrollment true "수강신청 정보"
// @Success 201 {object} map[string]string
// @Failure 400 {string} string "Bad Request"
// @Failure 500 {string} string "Internal Server Error"
// @Router /enrollments [post]
func (h *EnrollmentHandler) CreateEnrollment(w http.ResponseWriter, r *http.Request) {
	var enrollment models.Enrollment
	if err := json.NewDecoder(r.Body).Decode(&enrollment); err != nil {
		http.Error(w, "Invalid JSON", http.StatusBadRequest)
		return
	}

	err := h.Service.CreateEnrollment(&enrollment)
	if err != nil {
		log.Println("등록 등록 실패:", err)
		http.Error(w, "Failed to create enrollment", http.StatusInternalServerError)
		return
	}
	writeJSON(w, http.StatusCreated, map[string]string{"message": "Enrollment created"})
}

func (h *EnrollmentHandler) EnrollmentByIDHandler(w http.ResponseWriter, r *http.Request) {
	switch r.Method {
	case http.MethodGet:
		h.GetEnrollmentByID(w, r)
	case http.MethodPut:
		h.UpdateEnrollmentByID(w, r)
	case http.MethodDelete:
		h.DeleteEnrollmentByID(w, r)
	default:
		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
	}
}

// GetEnrollmentByID 특정 수강신청의 정보를 조회합니다
// @Summary 특정 수강신청 조회
// @Description 특정 수강신청의 정보를 조회합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Param enrollment_id path string true "수강신청 ID"
// @Success 200 {object} models.Enrollment
// @Failure 400 {string} string "Bad Request"
// @Failure 404 {string} string "Enrollment not found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /enrollments/{enrollment_id} [get]
func (h *EnrollmentHandler) GetEnrollmentByID(w http.ResponseWriter, r *http.Request) {
	enrollmentID := getIDFromPath(r.URL.Path)
	if enrollmentID == "" {
		http.Error(w, "Missing enrollment ID", http.StatusBadRequest)
		return
	}

	enrollment, err := h.Service.GetEnrollmentByID(enrollmentID)
	if err != nil {
		if err == sql.ErrNoRows {
			http.Error(w, "Enrollment not found", http.StatusNotFound)
		} else {
			log.Println("등록 조회 실패:", err)
			http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		}
		return
	}
	writeJSON(w, http.StatusOK, enrollment)
}

// UpdateEnrollmentByID 특정 수강신청의 정보를 수정합니다
// @Summary 수강신청 정보 수정
// @Description 특정 수강신청의 정보를 수정합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Param enrollment_id path string true "수강신청 ID"
// @Param enrollment body models.Enrollment true "수정할 수강신청 정보"
// @Success 200 {object} map[string]string
// @Failure 400 {string} string "Bad Request"
// @Failure 404 {string} string "Enrollment not found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /enrollments/{enrollment_id} [put]
func (h *EnrollmentHandler) UpdateEnrollmentByID(w http.ResponseWriter, r *http.Request) {
	enrollmentID := getIDFromPath(r.URL.Path)
	if enrollmentID == "" {
		http.Error(w, "Missing enrollment ID", http.StatusBadRequest)
		return
	}

	var enrollment models.Enrollment
	if err := json.NewDecoder(r.Body).Decode(&enrollment); err != nil {
		http.Error(w, "Invalid JSON", http.StatusBadRequest)
		return
	}

	err := h.Service.UpdateEnrollment(enrollmentID, &enrollment)
	if err != nil {
		if err == sql.ErrNoRows {
			http.Error(w, "Enrollment not found", http.StatusNotFound)
		} else {
			log.Println("등록 수정 실패:", err)
			http.Error(w, "Failed to update enrollment", http.StatusInternalServerError)
		}
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"message": "Enrollment updated"})
}

// DeleteEnrollmentByID 특정 수강신청을 삭제합니다
// @Summary 수강신청 삭제
// @Description 특정 수강신청을 삭제합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Param enrollment_id path string true "수강신청 ID"
// @Success 200 {object} map[string]string
// @Failure 400 {string} string "Bad Request"
// @Failure 404 {string} string "Enrollment not found"
// @Failure 500 {string} string "Internal Server Error"
// @Router /enrollments/{enrollment_id} [delete]
func (h *EnrollmentHandler) DeleteEnrollmentByID(w http.ResponseWriter, r *http.Request) {
	enrollmentID := getIDFromPath(r.URL.Path)
	if enrollmentID == "" {
		http.Error(w, "Missing enrollment ID", http.StatusBadRequest)
		return
	}

	err := h.Service.DeleteEnrollment(enrollmentID)
	if err != nil {
		if err == sql.ErrNoRows {
			http.Error(w, "Enrollment not found", http.StatusNotFound)
		} else {
			log.Println("등록 삭제 실패:", err)
			http.Error(w, "Failed to delete enrollment", http.StatusInternalServerError)
		}
		return
	}
	writeJSON(w, http.StatusOK, map[string]string{"message": "Enrollment deleted"})
}

// EnrollmentByStudentIDHandler 학생별 수강신청 목록을 조회합니다
// @Summary 학생별 수강신청 목록 조회
// @Description 특정 학생의 수강신청 목록을 조회합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Param student_id path string true "학생 ID"
// @Success 200 {array} models.Enrollment
// @Failure 400 {string} string "Bad Request"
// @Failure 500 {string} string "Internal Server Error"
// @Router /students/{student_id}/enrollments [get]
func (h *EnrollmentHandler) EnrollmentByStudentIDHandler(w http.ResponseWriter, r *http.Request) {
	studentID := getIDFromPath(r.URL.Path)
	if studentID == "" {
		http.Error(w, "Missing student ID", http.StatusBadRequest)
		return
	}

	enrollments, err := h.Service.GetEnrollmentsByStudentID(studentID)
	if err != nil {
		log.Println("등록 조회 실패:", err)
		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		return
	}
	writeJSON(w, http.StatusOK, enrollments)
}

// EnrollmentByClassIDHandler 수업별 수강신청 목록을 조회합니다
// @Summary 수업별 수강신청 목록 조회
// @Description 특정 수업의 수강신청 목록을 조회합니다
// @Tags enrollments
// @Accept json
// @Produce json
// @Param class_id path string true "수업 ID"
// @Success 200 {array} models.Enrollment
// @Failure 400 {string} string "Bad Request"
// @Failure 500 {string} string "Internal Server Error"
// @Router /classes/{class_id}/enrollments [get]
func (h *EnrollmentHandler) EnrollmentByClassIDHandler(w http.ResponseWriter, r *http.Request) {
	classID := getIDFromPath(r.URL.Path)
	if classID == "" {
		http.Error(w, "Missing class ID", http.StatusBadRequest)
		return
	}

	enrollments, err := h.Service.GetEnrollmentsByClassID(classID)
	if err != nil {
		log.Println("등록 조회 실패:", err)
		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
		return
	}
	writeJSON(w, http.StatusOK, enrollments)
}